+++ app/Http/Controllers/TrailsController.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const User = use('App/Model/User')
+const Hike = use('App/Model/Hike')
 const Trail = use('App/Model/Trail')
 
 class TrailsController {
@@ -7,15 +9,21 @@ class TrailsController {
 	* create (request, response){
 		let user = request.authUser;
 		let data = request.only('title', 'path', 'distance', 'max_elevation', 'min_elevation', 'img_url', 'description') // get new data
-		let exists = yield Trail.findBy('title', data.title) // check if title already exists in database
-		if (exists){
-			response.status(409).json({error: "Trail Name already taken!"})
-		} else {
-			data.user_id = user.id
-			data.username = user.username;
-			let trail = yield Trail.create(data)
-			response.status(201).json(trail)
-		}		
+
+		data.user_id = user.id
+		data.username = user.username;
+		let trail = yield Trail.create(data)
+
+		let trails_count = yield Trail.query().table('trails')
+			.countDistinct('id')
+			.where('user_id', user.id)
+
+		let update_user = yield User.findBy('id', user.id)
+		update_user.trails_created = trails_count[0].count;
+
+		yield update_user.save();
+
+		response.status(201).json(trail)		
 	}
 
 	* delete(request, response){
@@ -27,10 +35,38 @@ class TrailsController {
 		} else if (user.id !== trail.user_id){
 			response.status(403).json({error: "Trail does not belong to user"})
 		} else {
+
+
 			yield Trail.query().table('points')
 				.where('trail_id', trail_id)
 				.delete();
+
+			yield Trail.query().table('hikes')
+				.where('trail_id', trail_id)
+				.delete();
+
 			yield trail.delete();
+
+
+			let hikes_count = yield Hike.query().table('hikes')
+				.count('*')
+				.where('user_id', user.id);
+
+			let trails_count = yield Hike.query().table('hikes')
+				.countDistinct('trail_id')
+				.where('user_id', user.id)
+
+			let trails_created_count = yield Trail.query().table('trails')
+				.countDistinct('id')
+				.where('user_id', user.id)
+
+			let update_user = yield User.findBy('id', user.id)
+			update_user.hikes = hikes_count[0].count;
+			update_user.trails_hiked = trails_count[0].count;
+			update_user.trails_created = trails_created_count[0].count;
+
+			yield update_user.save();
+
 			response.status(204).send()
 		}
